Rename to gitter, so we won't conflict with import git
